Retails Sales/requirements.txt
pandas==1.5.3
numpy==1.23.4
matplotlib==3.6.2
streamlit==1.17.0
statsmodels==0.13.2
fbprophet==0.7.1
scikit-learn==1.0.2
